Cleaner implementation.
