+++ src/main/java/com/wjduquette/george/GameView.java
@@ -166,85 +166,6 @@ public class GameView extends StackPane {
         return region;
     }
 
-    /**
-     * Describes a feature, based on what it is.  Supported features include
-     * Signs and Mannikins.
-     * @param id The feature entity's ID
-     */
-    public void describeFeature(long id) {
-        repaint();
-        var entity = region.get(id);
-
-        if (entity.sign() != null) {
-            var key = entity.sign().key();
-            var text = region.getInfo(key, "text");
-
-            displayTextBlock(entity, text);
-        } else if (entity.mannikin() != null) {
-            var key = entity.mannikin().key();
-            StringBuilder buff = new StringBuilder();
-            buff.append(region.getInfo(key, "label")).append("\n\n");
-            buff.append(region.getInfo(key, "description")).append("\n\n");
-
-            List<String> greetings = region.info().values(key + ".greeting*");
-            buff.append(random.pickFrom(greetings));
-
-            displayTextBlock(entity, buff.toString());
-        }
-    }
-
-    public void displayTextBlock(Entity entity, String text) {
-        repaint();
-        System.out.println("GameView size=" + getWidth() + "x" + getHeight());
-        var panel = new SimplePanel(this, entity, text);
-        panel.setOnClose(() -> getChildren().remove(panel));
-        getChildren().add(panel);
-
-
-//        var xLeft = 50.0;
-//        var yTop = 50.0;
-//        var width = canvas.getWidth() - 100;
-//        var height= canvas.getHeight() - 100;
-//
-//        targets.clear();
-//        var box = new BoundingBox(0, 0, canvas.getWidth(), canvas.getHeight());
-//        var input = new UserInput.Continue();
-//        targets.add(new ClickTarget(box, () -> fireInputEvent(input)));
-//
-//        // The background
-//        canvas.gc().setFill(Color.DARKBLUE);
-//        canvas.gc().fillRect(xLeft, yTop, width, height);
-//
-//        // The image
-//        var ix = xLeft + 20;
-//        var iy = yTop + 40;
-//        var terrain = region.getTerrain(entity.cell());
-//        canvas.gc().drawImage(ImageUtils.embiggen(terrain.image(), 2), ix, iy);
-//        canvas.gc().drawImage(ImageUtils.embiggen(img(entity.sprite()), 2), ix, iy);
-//
-//        // The "click to continue"
-//        canvas.gc().setFill(Color.WHITE);
-//        canvas.gc().setFont(Font.font("Helvetica", 14));
-//        canvas.gc().fillText("Click to continue...",
-//            xLeft + 20,
-//            yTop + height - 20);
-//
-//        // The game text
-//        Text block = new Text(text);
-//        block.setFont(Font.font("Helvetica", 18));
-//        block.setWrappingWidth(width - 80);
-//        block.setFill(Color.WHITE);
-//        block.setOnMouseClicked(evt -> fireInputEvent(new UserInput.Continue()));
-//        StackPane.setAlignment(block, Pos.TOP_LEFT);
-//        StackPane.setMargin(block, new Insets(
-//            yTop + 40,            // Top
-//            xLeft + width - 20,   // Right
-//            yTop + height - 40,   // Bottom
-//            xLeft + 130));        // Left
-//
-//        getChildren().setAll(canvas, block);
-    }
-
     public void repaint() {
         canvas.clear();
         targets.clear();
