+++ src/main/java/com/wjduquette/george/widgets/SimplePanel.java
@@ -1,15 +1,11 @@
 package com.wjduquette.george.widgets;
 
 import com.wjduquette.george.App;
-import com.wjduquette.george.GameView;
 import com.wjduquette.george.ecs.Entity;
-import com.wjduquette.george.ecs.Sprite;
 import com.wjduquette.george.graphics.ImageUtils;
-import com.wjduquette.george.model.Region;
 import javafx.geometry.Insets;
 import javafx.geometry.Point2D;
 import javafx.geometry.VPos;
-import javafx.scene.image.Image;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.paint.Color;
 import javafx.scene.text.Font;
@@ -19,13 +15,14 @@ import javafx.scene.text.Font;
  */
 public class SimplePanel extends CanvasPane {
     private final static double INSET = 50;
-    private final GameView main;
+
+    private final App app;
     private final Entity entity;
     private final String text;
     private Runnable onClose = null;
 
-    public SimplePanel(GameView main, Entity entity, String text) {
-        this.main = main;
+    public SimplePanel(App app, Entity entity, String text) {
+        this.app = app;
         this.entity = entity;
         this.text = text;
         setPadding(new Insets(INSET));
@@ -47,7 +44,7 @@ public class SimplePanel extends CanvasPane {
     }
 
     private void repaint() {
-        var region = main.getRegion();
+        var region = app.getCurrentRegion();
         var w = getWidth() - 2*INSET;
         var h = getHeight() - 2*INSET;
 
@@ -59,7 +56,7 @@ public class SimplePanel extends CanvasPane {
         var ix = 30;
         var iy = 30;
         var terrain = region.getTerrain(entity.cell());
-        var sprite = main.getSprites().get(entity.sprite().name());
+        var sprite = app.sprites().get(entity.sprite().name());
         gc().drawImage(ImageUtils.embiggen(terrain.image(), 2), ix, iy);
         gc().drawImage(ImageUtils.embiggen(sprite, 2), ix, iy);
 
@@ -74,9 +71,5 @@ public class SimplePanel extends CanvasPane {
         gc().setFont(Font.font("Helvetica", 14));
         gc().setTextBaseline(VPos.BASELINE);
         gc().fillText("Click to continue...", 30, h - 30);
-    }
-
-    private Image img(Sprite sprite) {
-        return main.getSprites().get(sprite.name());
     }
 }
